Satellite upgrade needs the Satellite server repository enabled
Due to rubygem-foreman_maintain existing in the modular metadata for
Satellite repository but not the Maintenance repository, dnf will
not see a newer version as being upgradable. This has to be achieved
by the Satellite repository for the newer version being enabled
where the rubygem-foreman_maintain package is part of the
modular metadata.

This change is a work around to ensure user's have a working upgrade
experience while a more permenant solution is worked on around
modular metadata.

Ref: BZ#2143451